Refactor : files naming conv move compts to components folder
