+++ apps/buildproof_dapp/app/routes/app+/submit-project.tsx
@@ -1,4 +1,4 @@
-import { ProjectSubmission } from '../projectSubmission/ProjectSubmission'
+import { ProjectSubmission } from '../../components/project-submission/project-submission'
 
 export default function SubmitProject() {
   return <ProjectSubmission />
